Updated code to properly parse data and import it into the DB.
